trying to fix this part 2
