+++ src/RelayEnvironment.js
@@ -1,24 +1,11 @@
 import {Environment, Network, RecordSource, Store} from 'relay-runtime'
-import {AUTH_TOKEN} from './constants'
-async function fetchGraphQL(params, variables) {
-  const accessToken = localStorage.getItem(AUTH_TOKEN)
-  const headers = {'Content-Type': 'application/json'}
+import fetchGraphQL from './fetchGraphql'
 
-  if (accessToken) {
-    headers.Authorization = `Bearer ${accessToken}`
-  }
-
-  const response = await fetch('http://technify-api.test/graphql', {
-    body: JSON.stringify({
-      query: params.text,
-      variables,
-    }),
-    headers,
-    method: 'POST',
-  })
-  return response.json()
+async function fetchRelay(params, variables) {
+  return fetchGraphQL(params.text, variables)
 }
+
 export default new Environment({
-  network: Network.create(fetchGraphQL),
+  network: Network.create(fetchRelay),
   store: new Store(new RecordSource()),
 })
