+++ src/components/CategoriesList.jsx
@@ -18,6 +18,8 @@ export default function CategoriesList() {
                   node {
                     id
                     name
+                    description
+                    currentPrice
                   }
                 }
               }
