+++ src/pages/home/index.jsx
@@ -26,6 +26,7 @@ export default function Home() {
             node {
               name
               description
+              currentPrice
             }
           }
         }
